fix some printf()-like calls (format specifiers / argument types)

+++ src/engine/qcommon/vm.c
@@ -394,7 +394,7 @@ static void VM_InitSanity( vm_t *vm )
 }
 
 #ifndef NDEBUG
-#define VM_Insanity(c,n) Com_Error( ERR_DROP, "And it's a good night from vm. [%ld %s]", (c), (n) );
+#define VM_Insanity(c,n) Com_Error( ERR_DROP, "And it's a good night from vm. [%ld %s]", (long)(c), (n) );
 #else
 #define VM_Insanity(c,n) Com_Error( ERR_DROP, "And it's a good night from vm." );
 #endif
